Initial SDL3 port.
This work was done on YouTube: https://youtu.be/i3yVqWYFbCE

The SDL3 headers are in the commit before this one, so as not to junk up this
specific commit.

This needs further work; I mentioned a few things in the video. At a minimum,
the patch uses spaces instead of tabs, and that needs to be cleaned up.